moved notion api queue and guild pages to services folder and removed api folder.
